Refactor tools to use a single CLI entrypoint
Instead of having separate binaries for the various tools, use a single
CLI entrypoint "respdiff" and call the tools as subcommands, e.g.:

$ respdiff diff-answers ...

This makes it easier to distibute the tool and reduces the duplicity
around CLI argument parsing.

In addition to the CLI changes, the tools have been renamed:

msgdiff -> diff-answers
transceiver -> transceive

clap was also updated to the latest version and its declarative facility
was used for argument parsing.